transformation/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-monitor-opentelemetry
azure-functions
azure-eventhub
azure-cosmos
pydantic